feat: Implement group purchase which splits the amount among group members
May consider using Optional<String> as the element of return type for parse method in CommandParser, such that the return statement is more consistent. Or I will need to come up with something that can replace null to represent absent of a value as nmodifiable list does not allow storing null values.